refactor analytics alerts into modular handlers
