+++ src/lib/analytics/setup.ts
@@ -1,0 +1,8 @@
+if (process.env.SLACK_WEBHOOK_URL) {
+  await import('./alerts/slack');
+}
+
+if (process.env.EMAIL_WEBHOOK_URL) {
+  await import('./alerts/email');
+}
+
